+++ tests/TestClasses/Events/TestEvent.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Spatie\EventSourcing\Tests\TestClasses\Events;
+
+use Spatie\EventSourcing\StoredEvents\ShouldBeStored;
+
+class TestEvent extends ShouldBeStored
+{
+    public function __construct()
+    {
+    }
+}